code optimization - made more variables local
